# Improvements to the "Scale" SIMD algorithm