… test/Microsoft.ML.CpuMath.PerformanceTests/AvxPerformanceTests.cs
… test/Microsoft.ML.CpuMath.PerformanceTests/CpuMathNativeUtils.cs
… test/Microsoft.ML.CpuMath.PerformanceTests/NativePerformanceTests.cs
… test/Microsoft.ML.CpuMath.PerformanceTests/SsePerformanceTests.cs
… test/Microsoft.ML.CpuMath.UnitTests.netcoreapp/UnitTests.cs