@Anipik 2018-10-03
Results in some minor perf improvements for the ```Microsoft.ML.CpuMath.PerformanceTests```

```
Before
```
| Method       |     Mean |    Error  |   StdDev  |
|------------- |---------:|----------:|----------:|
| AvxScaleU    | 181.7 us | 3.400 us  | 3.339 us  |
| NativeScaleU | 172.1 us | 0.2690 us | 0.2385 us |
| SseScaleU    | 172.3 us | 0.5444 us | 0.4826 us |

```
After
```
| Method       |     Mean |     Error |    StdDev |
|--------------|---------:|----------:|----------:|
|  AvxScale    | 184.1 us | 0.9566 us | 0.7988 us |
|  NativeScale | 173.1 us | 1.082 us  | 0.8446 us |
|  AvxScale    | 172.6 us | 0.9265 us | 0.8213 us |