@danmosemsft 2018-10-05
@Anipik are before and after numbers backwards? They all got worse (albeit not necessarily significantly)\

Also are these perf tests of aligned or unaligned (or a random mixture)