@Anipik 2018-10-05
>  Also are these perf tests of aligned or unaligned (or a random mixture)

The data could be aligned or unaligned. The before algorithm was considering data to be always unaligned.

>  how good is the code coverage

I have tested all the code paths using breakpoints. But I will add more tests before the MatMul change because in that change there are lot of cases. No much cases in scale algorithm.

> Oh I see most of it is from Tanner's

Yes, tanners change acts as a blueprint. There will be always these four code paths. just the code and operations in them will slightly vary.

> are before and after numbers backwards? They all got worse (albeit not necessarily significantly)\

No the numbers are correct. yeah but they are within the error range