@Anipik 2018-10-08
@danmosemsft I updated the numbers, there is slight improvement in the numbers in most of the runs.